Refactor with new paths and utils modules
